Move cache to cache.c
